frlg rng fixes and tweaks
