+++ SerialPrograms/Source/PokemonFRLG/Programs/RngManipulation/PokemonFRLG_BlindNavigation.h
@@ -21,13 +21,20 @@ namespace PokemonFRLG{
         starters,
         magikarp,
         hitmon,
+        hitmonchan,
+        hitmonlee,
         eevee,
         lapras,
         fossils,
+        omanyte,
+        kabuto,
+        aerodactyl,
         gamecornerabra,
         gamecornerclefairy,
         gamecornerdratini,
         gamecornerbug,
+        gamecornerscyther,
+        gamecornerpinsir,
         gamecornerporygon,
         togepi,
         staticencounter,
