+++ SerialPrograms/Source/PokemonFRLG/Programs/RngManipulation/PokemonFRLG_RngDisplays.h
@@ -10,6 +10,7 @@
 #include <vector>
 #include "Common/Cpp/Options/StringOption.h"
 #include "Common/Cpp/Options/TextEditOption.h"
+#include "Common/Cpp/Options/FloatingPointOption.h"
 #include "CommonFramework/Notifications/EventNotificationsTable.h"
 #include "NintendoSwitch/NintendoSwitch_SingleSwitchProgram.h"
 #include "Pokemon/Pokemon_AdvRng.h"
@@ -59,12 +60,22 @@ public:
 };
 
 
-class PossibleHitsDisplay : public GroupOption{
+class RngCalibrationDisplay : public GroupOption{
 public:
-    PossibleHitsDisplay();
+    RngCalibrationDisplay();
 
-    void set(const std::vector<AdvRngState>& rng_states);
-    void set(const std::map<AdvRngState, AdvPokemonResult>& hits_map);
+    void set(
+        double s_calibraiton, 
+        double c_calibration, 
+        double a_calibration, 
+        std::vector<AdvRngState>& rng_states
+    );
+    void set(
+        double s_calibration, 
+        double c_calibration, 
+        double a_calibration, 
+        const std::map<AdvRngState, AdvPokemonResult>& hits_map
+    );
     void reset();
 
 private:
@@ -72,6 +83,9 @@ private:
     static std::string get_hits_string(const std::vector<AdvRngState>& rng_states);
     static std::string get_hits_string(const std::map<AdvRngState, AdvPokemonResult>& hits_map);
 public:
+    FloatingPointOption seed_calibration;
+    FloatingPointOption csf_calibration;
+    FloatingPointOption advances_calibration;
     StringOption hits;
 };
